+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * **noncore**: Add POST /users endpoint
 * **noncore**: Endpoint for getting guardians by project id
+* **noncore**: Update is_updatable value in the Guardians table
 
 ### Performance improvements
 * **noncore**: Delete code and table related to FilterPresets, Reports with Attachments, UserLocations, UserTokens, ResetPasswordTokens
@@ -11,6 +12,7 @@
 
 ### Bug Fixes
 * **core**: Set arbimon ingestion request timeout to 10 seconds
+* **noncore**: Fixed the issue with updating guardians data
 
 ## 1.1.3 (2022-04-05)